Extract direct-access station key into hook
Move shared state and click/visual logic for direct-access station keys into a new useDirectAccessStationKey hook. Update DirectAccessStationKey to consume the hook (reducing its internal logic and imports) and add GeoPageDirectButton/GeoPageButton changes so GeoPage can reuse the same hook for direct station buttons. Adds vacs-client/frontend/src/hooks/use-direct-access-station-key.ts and updates DirectAccessStationKey.tsx and GeoPage.tsx to use the hook
